+++ apps/client/src/routes/+page.svelte
@@ -1,31 +1,20 @@
 <script lang="ts">
 import Dashboard from "$components/dashboard-summary.svelte";
-import type { DashboardSummary } from "$lib/dashboard/types.js";
-import { EVENT_DASHBOARD_SUMMARY_UPDATE } from "$lib/events.js";
 import { ShieldEllipsis } from "lucide-svelte";
 import { derived } from "svelte/store";
-import { source } from "sveltekit-sse";
-
-const eventSource = source("/events/dashboard-summary");
-const channel = eventSource.select(EVENT_DASHBOARD_SUMMARY_UPDATE);
-const summary = channel.transform((value: string) => {
-	const d = JSON.parse(value) as DashboardSummary;
-	return {
-		pullRequests: d.pullRequests ?? [],
-		securityAdvisories: d.securityAdvisories ?? [],
-		products: d.products ?? [],
-	};
-});
+import { summaryStore } from "$lib/dashboard/store";
 
 const pullRequests = derived(
-	summary,
+	summaryStore,
 	($summary?) => $summary?.pullRequests ?? [],
 );
+
 const securityAdvisories = derived(
-	summary,
+	summaryStore,
 	($summary?) => $summary?.securityAdvisories ?? [],
 );
-const products = derived(summary, ($summary?) => $summary?.products ?? []);
+
+const products = derived(summaryStore, ($summary?) => $summary?.products ?? []);
 </script>
 
 <h1 class="heading-1">Dashboard</h1>
@@ -34,7 +23,7 @@ const products = derived(summary, ($summary?) => $summary?.products ?? []);
     <a class="text-xs text-link" href="/products/create">Add additional products</a>
 </div>
 
-{#if $summary === null}
+{#if $summaryStore === null}
     <div class="w-full my-44 flex flex-col items-center justify-center">
         <ShieldEllipsis size="150" />
         <h3 class="heading-3">Loading...</h3>
